requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.6.1
alembic==1.13.1
httpx==0.27.0
